+++ tests/Flawright.E2ETests.TestApp/App.xaml.cs
@@ -1,10 +1,8 @@
-using System.Windows;
-
 namespace Flawright.E2ETests.TestApp;
 
 /// <summary>
 /// Application entry point for the Flawright E2E test target.
 /// </summary>
-public partial class App : Application
+public partial class App : System.Windows.Application
 {
 }